+++ CHANGELOG.md
@@ -11,11 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add support for upcoming `<<<>>>` syntax.
 - Improve syntax highlighting for global variables.
+- Diagnostics: Add support for spotting variables that are defined but never used.
 
 ### Fixed
 
 ### Changed
 
+- `extension.ts`: Clean up file by breaking nested function up into separate files.
 - Include `iterationmode` in `for` loop intellisense.
 
 ### Removed